[FAB-16559] Support Go plugins in fabric-cli
This patch adds support for Go plugins as fabric-cli plugins. A Go
plugin implements a factory that returns a cobra.Command. The returned
command is added as a sub-command of the main fabric command.

At load time, the binary specified in the plugin.yaml file is first
loaded as a Go plugin. If this fails then it reverts to the current
behavior.

Signed-off-by: Bob Stasyszyn <Bob.Stasyszyn@securekey.com>
Change-Id: I27bb663c4ece72a81cbbf277ced1f0e92e3d1033